Adding Core library
