+++ Package.swift
@@ -9,6 +9,7 @@ let package = Package(
         .macOS(.v10_15)
     ],
     products: [
+        .library(name: "Core", targets: ["Core"]),
         .executable(name: "macdb", targets: ["macdb"]),
     ],
     dependencies: [
@@ -16,11 +17,12 @@ let package = Package(
         // .package(url: /* package url */, from: "1.0.0"),
     ],
     targets: [
-        // Targets are the basic building blocks of a package. A target can define a module or a test suite.
-        // Targets can depend on other targets in this package, and on products in packages which this package depends on.
+        .target(
+            name: "Core",
+            dependencies: []),
         .target(
             name: "macdb",
-            dependencies: []),
+            dependencies: ["Core"]),
         .testTarget(
             name: "macdbTests",
             dependencies: ["macdb"]),
